requirements.txt
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
h11==0.12.0
idna==3.3
outcome==1.1.0
pycparser==2.21
pylint==2.7.4
pyOpenSSL==21.0.0
PyYAML==6.0
requests==2.27.0
robotframework==4.1.3
robotframework-jsonlibrary==0.3.1
robotframework-pabot==2.1.0
robotframework-pythonlibcore==3.0.0
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==5.1.3
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
wsproto==1.0.0
#tk==0.1.0
